Add initial working version of AST
Most parts of the recursive descent are updated to build corresponding
AST nodes. The AST nodes themselves are mainly completed, with print and
codegen method implemented.

A function and procedure declaration has yet not been translated into
the AST, print and codegen methods have yet to be implemented.
The code is also missing a scope handling, operation with the symbol
table and several other stuff.